# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShadowID Forensic Core API",
    "version": "1.4.0",
    "description": "Production-grade modular monolith backend for ShadowID — Digital footprint exposure intelligence, impersonation detection, and document defense platform engineered for India's digital identity ecosystem. Team GIGABYTE (Anshul, Tanishq, Himank), Build With Bharat 3.0, Chitkara University, Himachal Pradesh."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.shadowid.in",
      "description": "Production API (India Region)"
    }
  ],
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "summary": "Get Current Authenticated Profile and Active Workspace",
        "operationId": "getCurrentUser",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile with workspace context",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "summary": "List User Workspaces",
        "operationId": "listWorkspaces",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Workspace" }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comparisons": {
      "post": {
        "summary": "Perform Multi-Signal Impersonation Comparison",
        "operationId": "compareProfiles",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProfileComparisonRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comparison metrics and verdict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileComparisonData" }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/analyze": {
      "post": {
        "summary": "Run Tesseract eng/hin OCR and OpenCV Morphological Checks",
        "operationId": "analyzeDocument",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentAnalysisRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OCR bounding boxes and anomaly findings",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DocumentAnalysisData" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/billing/orders": {
      "post": {
        "summary": "Create Razorpay Order for 30-Day Pro Pass (₹499)",
        "operationId": "createBillingOrder",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Razorpay order payload (amount in paise)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BillingOrderResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/verify": {
      "post": {
        "summary": "Verify Razorpay Payment Signature and Grant Pro Entitlement",
        "operationId": "verifyPayment",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyPaymentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment confirmed and Pro pass activated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentVerificationResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/webhooks/razorpay": {
      "post": {
        "summary": "Razorpay Server Webhook Endpoint",
        "operationId": "handleRazorpayWebhook",
        "parameters": [
          { "name": "X-Razorpay-Signature", "in": "header", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Webhook processed" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "fullName": { "type": "string" },
          "workspaceId": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["owner", "analyst", "viewer"] },
          "isPro": { "type": "boolean" }
        },
        "required": ["id", "email", "fullName", "workspaceId", "role", "isPro"]
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "tier": { "type": "string", "enum": ["free", "pro", "enterprise"] },
          "retentionDays": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "tier", "retentionDays", "createdAt"]
      },
...
      "ProfileComparisonRequest": {
        "type": "object",
        "properties": {
          "referenceSubjectId": { "type": "string" },
          "candidateUrl": { "type": "string" },
          "candidateName": { "type": "string" },
          "candidateHandle": { "type": "string" },
          "candidateBio": { "type": "string" }
        },
        "required": ["referenceSubjectId", "candidateName", "candidateHandle"]
      },
      "ProfileComparisonData": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "referenceSubject": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "handle": { "type": "string" },
              "bio": { "type": "string" },
              "institution": { "type": "string" },
              "city": { "type": "string" },
              "avatarUrl": { "type": "string" }
            },
            "required": ["name", "handle"]
          },
          "candidateProfile": {
            "type": "object",
            "properties": {
              "platform": { "type": "string" },
              "name": { "type": "string" },
              "handle": { "type": "string" },
              "bio": { "type": "string" },
              "location": { "type": "string" }
            },
            "required": ["platform", "name", "handle"]
          },
          "metrics": {
            "type": "object",
            "properties": {
              "nameTokenOverlap": { "type": "integer" },
              "handleSimilarity": { "type": "integer" },
              "bioTextOverlap": { "type": "integer" },
              "avatarPerceptualMatch": { "type": "integer" },
              "contextualAgreement": { "type": "integer" }
            },
            "required": ["nameTokenOverlap", "handleSimilarity", "bioTextOverlap"]
          },
          "contradictions": { "type": "array", "items": { "type": "string" } },
          "verdict": { "type": "string", "enum": ["Limited overlap", "Similarities found", "Needs review", "Insufficient evidence"] },
          "verdictNotes": { "type": "string" },
          "reviewerDecision": { "type": "string", "enum": ["dismiss", "flag_for_takedown", "monitor"], "nullable": true }
        },
        "required": ["id", "referenceSubject", "candidateProfile", "metrics", "verdict"]
      },
      "DocumentAnalysisRequest": {
        "type": "object",
        "properties": {
          "fileId": { "type": "string" },
          "category": { "type": "string", "enum": ["Aadhaar", "PAN", "Passport", "Generic ID"] }
        },
        "required": ["fileId", "category"]
      },
      "DocumentAnalysisData": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "documentCategory": { "type": "string" },
          "sampleLabel": { "type": "string" },
          "qualityScore": { "type": "integer" },
          "languageDetected": { "type": "string" },
          "verdict": { "type": "string", "enum": ["Requires review", "No configured inconsistencies found", "Insufficient image quality"] },
          "summary": { "type": "string" },
          "boxes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "label": { "type": "string" },
                "value": { "type": "string" },
                "maskedValue": { "type": "string" },
                "confidence": { "type": "integer" },
                "isAnomaly": { "type": "boolean" },
                "anomalyReason": { "type": "string", "nullable": true }
              },
              "required": ["id", "label", "value", "maskedValue", "confidence", "isAnomaly"]
            }
          },
          "inconsistencies": { "type": "array", "items": { "type": "string" } },
          "isHumanVerified": { "type": "boolean" }
        },
        "required": ["id", "documentCategory", "qualityScore", "verdict", "summary", "boxes", "isHumanVerified"]
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "plan": { "type": "string", "enum": ["pro_pass_30d"] }
        },
        "required": ["plan"]
      },
      "BillingOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string" },
          "amountPaise": { "type": "integer", "example": 49900 },
          "currency": { "type": "string", "example": "INR" },
          "keyId": { "type": "string" },
          "productName": { "type": "string", "example": "ShadowID 30-Day Pro Pass" }
        },
        "required": ["orderId", "amountPaise", "currency", "keyId", "productName"]
      },
      "VerifyPaymentRequest": {
        "type": "object",
        "properties": {
          "razorpayOrderId": { "type": "string" },
          "razorpayPaymentId": { "type": "string" },
          "razorpaySignature": { "type": "string" }
        },
        "required": ["razorpayOrderId", "razorpayPaymentId", "razorpaySignature"]
      },
      "PaymentVerificationResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "entitlement": { "type": "string" },
          "validUntilIST": { "type": "string" }
        },
        "required": ["success", "entitlement", "validUntilIST"]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication token missing, invalid, or expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": { "type": "string", "example": "AUTH_UNAUTHORIZED" },
                "message": { "type": "string", "example": "Missing or expired JWT session token" },
                "requestId": { "type": "string" }
              },
              "required": ["code", "message", "requestId"]
            }
          }
        }
      },
...
    }
  }
}